RDKB-35146 : Coverity fix for ccsp-cmn-lib
Reason for change: Due to Coverity fix in ccsp-cmn-lib
Test Procedure: Build through Jenkins
Risks: Low


RDKB-35360 : Coverity fix for ccsp-cmn-lib part2

Reason for change: Due to Coverity fix in ccsp-cmn-lib
Test Procedure: Build through Jenkins
Risks: Low


RDKB-35470: Coverity fix for ccsp-cmn-lib

Reason for change: Due to Coverity fix in ccsp-cmn-lib part 3
Test Procedure: Build through Jenkins
Risks: Low


RDKB-35519: Coverity fix for ccsp-cmn-lib

Reason for change: Due to Coverity fix in ccsp-cmn-lib part 4
Test Procedure: Build through Jenkins
Risks: Low

+++ source/util_api/web/components/WebRooFsm/web_roofsm_do_stuff.c
@@ -247,8 +247,9 @@ WebRooFsmDoGet
      * produced data which shall be returned as the entity in the response and not the source text
      * of the process, unless that text happens to be the output of the process.
      */
-    ulFileSize = AnscGetFileSize(hMappedFile);
-
+    /*CID :61576 Argument cannot be negative - Added negative check in def part*/
+    ulFileSize = (ULONG) AnscGetFileSize(hMappedFile);
+        
     if ( ulFileSize == 0 )
     {
         return  ANSC_STATUS_SUCCESS;
